requirements.txt
flask==2.3.3
Werkzeug==2.3.7
jinja2==3.1.2
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
webdriver-manager==4.0.0
pdfplumber==0.10.2
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.25.2
ollama==0.1.4
charset-normalizer==3.2.0
certifi==2023.7.22
idna==3.4
urllib3==2.0.4
soupsieve==2.5
trio==0.22.2
trio-websocket==0.10.4
wsproto==1.2.0
pypdfium2==4.19.0
importlib-metadata==6.8.0
cryptography==41.0.3
decorator==5.1.1
packaging==23.1
pyproject_hooks==1.0.0
sniffio==1.3.0
exceptiongroup==1.1.3
attrs==23.1.0
outcome==1.2.0
sortedcontainers==2.4.0
asyncio==3.4.3
pycparser==2.21
cffi==1.15.1
python-docx==0.8.11 